requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
streamlit>=1.38.0
pydantic>=2.8.0
python-dotenv>=1.0.1
openai>=1.40.0
openai-agents
httpx>=0.27.0
pymongo>=4.8.0
playwright>=1.46.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-mongodb>=0.2.0
langchain-text-splitters>=0.3.0
pypdf>=4.3.0
